requirements.txt
elasticsearch==8.11.1
pytz==2023.3.post1
pyyaml==6.0.1
requests==2.31.0
